Update README.md
Simple React Calculator


A basic and intuitive arithmetic calculator built with React, allowing users to perform simple addition and subtraction operations.


Features


- Perform basic arithmetic operations (+/- 1, 2, 3)
- Clear function to reset the result
- Simple and intuitive user interface


About


This calculator is designed to provide a straightforward and user-friendly experience. Built with React, JavaScript, and CSS, it demonstrates a basic implementation of a calculator.
→Simple React Calculator🔢

A basic and intuitive arithmetic calculator built with React, allowing users to perform simple addition and subtraction operations.

→Features

- Perform basic arithmetic operations (+/- 1, 2, 3)
- Clear function to reset the result
- Simple and intuitive user interface

About

This calculator is designed to provide a straightforward and user-friendly experience. Built with React, JavaScript, and CSS, it demonstrates a basic implementation of a calculator.